LayoutTests/pointer-lock/lock-element-not-in-dom.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../http/tests/resources/js-test-pre.js"></script>
<script src="../http/tests/resources/pointer-lock/pointer-lock-test-harness.js"></script>
</head>
<body>
<div>
  <div id="target1"></div>
</div>
<script>
    description("Test locking an element not in a document is rejected and pointerlockerror event dispatched.")
    window.jsTestIsAsync = true;

    targetDiv1 = document.getElementById("target1");

    todo = [
        function () {
            shouldNotBe("targetDiv1.parentElement", "null");
            expectOnlyErrorEvent("Remove targetDiv1 from document, and try to lock it.");
            targetDiv1.parentElement.removeChild(targetDiv1);
            shouldBe("targetDiv1.parentElement", "null");
            targetDiv1.webkitRequestPointerLock();
            // doNextStep called by event handler.
        },
    ];
    doNextStepWithUserGesture();
</script>
<script src="../http/tests/resources/js-test-post.js"></script>
</body>
</html>
