third_party/blink/web_tests/platform/generic/pointer-lock/lock-element-not-in-dom-expected.txt

Test locking an element not in a document is rejected and pointerlockerror event dispatched.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS targetDiv1.parentElement is not null
     Remove targetDiv1 from document, and try to lock it.
PASS targetDiv1.parentElement is null
PASS onpointerlockerror received after: Remove targetDiv1 from document, and try to lock it.
PASS successfullyParsed is true

TEST COMPLETE
